% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celsius_to_kelvin.R
\name{celcius_to_kelvin}
\alias{celcius_to_kelvin}
\title{Temperature Converters}
\usage{
celcius_to_kelvin(temp_celcius)
}
\arguments{
\item{temp_celcius}{A temperature in celcius}
}
\value{
Returns the value of the temperature in kelvin
}
\description{
Converts a temperature in celcius to kelvin
}
\details{
#' \code{celcius_to_kelvin} Converts a temperature in celcius to kelvin
}
\examples{
Pass a temperature in celcius in and it will return that temperature in kelvin
celcius_to_kelvin(5)



}
