% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fahrenheit_to_celcius.R
\name{fahrenheit_to_celcius}
\alias{fahrenheit_to_celcius}
\title{Temperature Converters}
\usage{
fahrenheit_to_celcius(temp_fahrenheit)
}
\arguments{
\item{temp_fahrenheit}{A temperature in fahrenheit}
}
\value{
Returns the value of the temperature in Celcius
}
\description{
Converts temperature
}
\details{
\code{fahrenheit_to_celcius} Converts a temperature in fahrenheit to celcius
}
\examples{
Pass a temperature in fahrenheit in and it will return that temperature in celcius
fahrenheit_to_celcius(5)


}
