% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fahrenheit_to_kelvin.R
\name{fahrenheit_to_kelvin}
\alias{fahrenheit_to_kelvin}
\title{Temperature Converters}
\usage{
fahrenheit_to_kelvin(temp_fahrenheit)
}
\arguments{
\item{temp_kelvin}{A temperature in kelvin}
}
\value{
Returns the value of the temperature in kelvin
}
\description{
Converts a temperature in fahrenheit to kelvin
}
\details{
#' \code{fahrenheit_to_kelvin} Converts a temperature in fahrenheit to kelvin
}
\examples{
Pass a temperature in fahrenheit in and it will return that temperature in kelvin
fahrenheit_to_kelvin(5)


}
