% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqft_to_sqmeters.R
\name{sqft_to_sqmeters}
\alias{sqft_to_sqmeters}
\title{converts area}
\usage{
sqft_to_sqmeters(sq_ft)
}
\arguments{
\item{sq_ft}{An area in sq_ft}
}
\value{
Returns an area in sqmeters
}
\description{
Converts an area in sqft to sqmeters
}
\details{
\code{Sqft_to_sqmeters} Converts an area in sqft to sqmeters
}
\examples{
Pass in an area in sqft and return an area in sqmeters
sqft_to_sqmeters(5)

}
