% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqft_to_sqmile.R
\name{sqft_to_sqmile}
\alias{sqft_to_sqmile}
\title{converts area}
\usage{
sqft_to_sqmile(sq_mile)
}
\arguments{
\item{sq_ft}{An area in sq_ft}
}
\value{
Returns an area in sqmiles
}
\description{
Converts an area in sqft to sqmiles
}
\details{
\code{Sqft_to_sqmiles} Converts an area in sqft to sqmiles
}
\examples{
Pass in an area in sqft and return an area in sqmiles
sqft_to_sqmile(5)

}
